// src/app/model/interfaces/Fornecedores.ts
export interface Fornecedores {

  id: number;
  cpfCnpj: string;
  nomeRazaoSocial: string;
  telefone: number;
  contato: string;
  email: string;
  rua: string;
  numero: string;
  bairro: string;
  cidade: string;
  estado: string;
  complemento: string;
  cep: string;
  inscEstadual: string;
  inscMunicipal: string;
}